requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
albumentations=1.3.1=pypi_0
blas=1.0=mkl
bottleneck=1.3.5=py38h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotli-python=1.0.9=py38h6a678d5_7
ca-certificates=2023.12.12=h06a4308_0
certifi=2023.11.17=py38h06a4308_0
cffi=1.16.0=py38h5eee18b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
contourpy=1.0.5=py38hdb19cb5_0
cryptography=41.0.3=py38h130f0dd_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h9c0eb46_1
dbus=1.13.18=hb2f20db_0
efficientnet-pytorch=0.7.1=pypi_0
expat=2.5.0=h6a678d5_0
filelock=3.13.1=py38h06a4308_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.12.2=pypi_0
giflib=5.2.1=h5eee18b_3
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h5py=3.10.0=pypi_0
icu=73.1=h6a678d5_0
idna=3.4=py38h06a4308_0
imageio=2.33.1=pypi_0
importlib_resources=6.1.1=py38h06a4308_1
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.2=py38h06a4308_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.4=py38h6a678d5_0
krb5=1.20.1=h568e23c_1
lazy-loader=0.3=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.8.1.2=0
libcups=2.4.2=ha637b67_0
libcurand=10.3.4.107=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libllvm14=14.0.6=hdb19cb5_3
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libpq=12.15=h37d81fd_1
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hf1b16e4_1
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.3=py38h5eee18b_0
matplotlib=3.7.2=py38h06a4308_0
matplotlib-base=3.7.2=py38h1128e8f_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py38h06a4308_0
munkres=1.1.4=py_0
mysql=5.7.24=he378463_2
ncurses=6.4=h6a678d5_0
networkx=3.1=py38h06a4308_0
numexpr=2.8.4=py38hc78ab66_1
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
opencv-python=4.9.0.80=pypi_0
opencv-python-headless=4.9.0.80=pypi_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1w=h7f8727e_0
packaging=23.1=py38h06a4308_0
pandas=2.0.3=py38h1128e8f_0
pcre=8.45=h295c915_0
pillow=10.0.1=py38ha6cbd5a_0
pip=23.3.1=py38h06a4308_0
platformdirs=3.10.0=py38h06a4308_0
ply=3.11=py38_0
pooch=1.7.0=py38h06a4308_0
protobuf=4.25.2=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py38h06a4308_0
pyparsing=3.0.9=py38h06a4308_0
pyqt=5.15.10=py38h6a678d5_0
pyqt5-sip=12.13.0=py38h5eee18b_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.15=h7a1cb2a_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.1.2=py3.8_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py38h06a4308_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=py38h5eee18b_0
qt-main=5.15.2=h110a718_10
qudida=0.0.4=pypi_0
readline=8.2=h5eee18b_0
requests=2.31.0=py38h06a4308_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=py38hf6e8229_1
setuptools=68.2.2=py38h06a4308_0
sip=6.7.12=py38h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
sympy=1.12=py38h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorboardx=2.6.2.2=pypi_0
threadpoolctl=3.2.0=pypi_0
tifffile=2023.7.10=pypi_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=py38h06a4308_0
torchtriton=2.1.0=py38
torchvision=0.15.2=cpu_py38h83e0c9b_0
tornado=6.3.3=py38h5eee18b_0
tqdm=4.65.0=py38hb070fc8_0
typing_extensions=4.9.0=py38h06a4308_0
urllib3=1.26.18=py38h06a4308_0
wheel=0.41.2=py38h06a4308_0
xz=5.4.5=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zipp=3.17.0=py38h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
